+++ src/main/java/controller/Controller.java
@@ -1,0 +1,4 @@
+package controller;
+
+public class Controller {
+}
